moved lucky number module and test to correct folders
