--- src/pykeafortune/luckyNumber.py
@@ -1,1 +1,0 @@
-# Lucky number file
